include the CAN timestamp in the can::Message structure
This timestamp is closer to the time that the message actually reached
the CAN card. This data therefore allows us to make sure when the actual
CAN packet has been received (in relation to the other CAN packets), and
ultimately will allow us to estimate the latency variability in-between
the card and the software.